+++ app/src/main/java/huette/kasse/activities/AddDrink.kt
@@ -1,4 +1,4 @@
-package huette.kasse
+package huette.kasse.activities
 
 import android.os.Bundle
 import android.widget.Button
@@ -7,6 +7,9 @@ import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import huette.kasse.DrinksAdapter
+import huette.kasse.R
+import huette.kasse.Variables
 
 class AddDrink : AppCompatActivity(), DrinksAdapter.OnItemClickListener {
     override fun onCreate(savedInstanceState: Bundle?) {
