+++ app/src/main/java/huette/kasse/activities/DeleteUser.kt
@@ -1,10 +1,13 @@
-package huette.kasse
+package huette.kasse.activities
 
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import huette.kasse.NamesAdapter
+import huette.kasse.R
+import huette.kasse.Variables
 
 class DeleteUser : AppCompatActivity(), NamesAdapter.OnItemClickListener {
     override fun onCreate(savedInstanceState: Bundle?) {
