+++ app/src/main/java/huette/kasse/activities/MainActivity.kt
@@ -1,4 +1,4 @@
-package huette.kasse
+package huette.kasse.activities
 
 import android.content.Intent
 import android.os.Bundle
@@ -9,6 +9,9 @@ import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import huette.kasse.NamesAdapter
+import huette.kasse.R
+import huette.kasse.Variables
 import huette.kasse.data.UserViewModel
 
 class MainActivity : AppCompatActivity(), NamesAdapter.OnItemClickListener {
