+++ app/src/main/java/huette/kasse/activities/Pay.kt
@@ -1,4 +1,4 @@
-package huette.kasse
+package huette.kasse.activities
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,6 +6,9 @@ import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import huette.kasse.NamesAdapter
+import huette.kasse.R
+import huette.kasse.Variables
 
 class Pay : AppCompatActivity(), NamesAdapter.OnItemClickListener {
     override fun onCreate(savedInstanceState: Bundle?) {
