+++ app/src/main/java/huette/kasse/activities/Uebersicht.kt
@@ -1,7 +1,8 @@
-package huette.kasse
+package huette.kasse.activities
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import huette.kasse.R
 
 class Uebersicht : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
