+++ app/src/main/java/huette/kasse/data/daos/UserDrinksDao.kt
@@ -11,9 +11,9 @@ import huette.kasse.data.entities.UserDrinks
 interface UserDrinksDao {
 
     @Insert(onConflict = OnConflictStrategy.IGNORE)
-    suspend fun addDrinkToUser(userDrink: UserDrinks)
+    fun addDrinkToUser(userDrink: UserDrinks)
 
-    @Query("SELECT user_id, drink_id, firstname, lastname, amount FROM user_drinks, users WHERE user_drinks.user_id = users.id ORDER BY firstName, lastName ASC")
+    @Query("SELECT user_id, drink_id, firstname, lastname, paid FROM user_drinks, users WHERE user_drinks.user_id = users.id ORDER BY firstName, lastName ASC")
     fun getAllData(): LiveData<List<UserDrinks>>
 
 }
