+++ app/src/main/java/huette/kasse/data/entities/UserDrinks.kt
@@ -4,11 +4,17 @@ import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.ForeignKey.CASCADE
+import java.sql.Date
+import java.sql.Time
+import java.sql.Timestamp
 
 @Entity(tableName = "user_drinks", primaryKeys = ["user_id", "drink_id"], foreignKeys = arrayOf(
     ForeignKey(
         entity = User::class, parentColumns = arrayOf("id"), childColumns = arrayOf("user_id"), onDelete = CASCADE
-    ), ForeignKey(entity = Drink::class, parentColumns = arrayOf("id"), childColumns = arrayOf("drink_id"), onDelete = CASCADE)
+    ),
+    ForeignKey(
+        entity = Drink::class, parentColumns = arrayOf("id"), childColumns = arrayOf("drink_id"), onDelete = CASCADE
+    )
 )
 )
 data class UserDrinks(
@@ -16,6 +22,10 @@ data class UserDrinks(
     val userID: Int,
     @ColumnInfo(name = "drink_id")
     val drinkID: Int,
-    @ColumnInfo(name = "amount")
-    val amount: Int,
-)
+    @ColumnInfo(name = "paid")
+    val paid: Boolean,
+    @ColumnInfo(name = "timestamp")
+    val timeStamp: Long,
+){
+    constructor(user_id: Int, drink_id: Int) : this(user_id, drink_id, false, System.currentTimeMillis())
+}
