+++ build.gradle
@@ -9,7 +9,7 @@ buildscript {
         classpath 'com.android.tools.build:gradle:4.2.2'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
 
-        classpath "androidx.navigation:navigation-safe-args-gradle-plugin:2.3.0-rc01"
+        classpath "androidx.navigation:navigation-safe-args-gradle-plugin:2.4.0-alpha05"
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
